issue #356 clean up and refactor
