+++ server/catgenome/src/main/java/com/epam/catgenome/manager/bed/parser/NggbMultiFormatBedFeature.java
@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2016 EPAM Systems
+ * Copyright (c) 2016-2021 EPAM Systems
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -26,7 +26,7 @@ package com.epam.catgenome.manager.bed.parser;
 
 import htsjdk.tribble.annotation.Strand;
 
-import java.awt.*;
+import java.awt.Color;
 import java.util.HashMap;
 import java.util.Map;
